+++ src/controllers/controller_Clientes.js
@@ -5,7 +5,6 @@ exports.create = async (req, res) => {
   const {
     nome,
     cnpj,
-    cnpjApto,
     razaoSocial,
     endereco,
     numero,
@@ -15,37 +14,40 @@ exports.create = async (req, res) => {
     emailResponsavel,
     nomeResponsavel,
     telefoneResponsavel,
-    Situacao
+    situacao
   } = req.body;
+
+  if( !nome || 
+      !cnpj ||
+      !razaoSocial || 
+      !emailResponsavel ||
+      !nomeResponsavel || 
+      !situacao){
+        return res.status(400).json({message: "Dados Inválidos"})
+  }
 
   try{
     const newClient = await prisma.client.create({
-    data:{
-      nome,
-      cnpj,
-      cnpjApto,
-      razaoSocial,
-      endereco,
-      numero,
-      complemento,
-      bairro,
-      cidade,
-      emailResponsavel,
-      nomeResponsavel,
-      telefoneResponsavel,
-      Situacao
-    }
-  })
-  console.log(newClient);
-  res.json({ status: 200, message: "Cliente cadastrado com sucesso!", newClient});
+      data:{
+        nome,
+        cnpj,
+        razaoSocial,
+        endereco,
+        numero,
+        complemento,
+        bairro,
+        cidade,
+        emailResponsavel,
+        nomeResponsavel,
+        telefoneResponsavel,
+        situacao
+      }
+    })
+    res.status(200).json({message: "Cliente cadastrado com sucesso!", newClient});
 
   } catch (err) {
     console.error(err);
-    res.json({
-      status: 400,
-      error: "Dados inválidos",
-      message: "Um ou mais campos estão inválidos",
-    });
+    res.status(500).end();
   }
 };
 
@@ -53,20 +55,24 @@ exports.searchAll = async (req, res) => {
 
   try{
     const clients = await prisma.client.findMany();
-    res.json({status: 200, clients})
+    if(clients.length < 1){
+      return res.status(404).json({message: "Dados Não Encontrados"})
+    }
+    res.status(200).json({clients})
+    
   } catch (err) {
     console.error(err);
-    res.json({
-      status: 400,
-      error: "Dados não encontrados",
-    });
+    res.status(500).end();
   }
 };
 
 exports.search = async (req, res) => {
   const {id} = req.params;
+  if(!id){
+    return res.status(400).json({message: "Dados Inválidos"})
+  }
+
   const id_ = parseInt(id);
-  
   try{
     const client = await prisma.client.findMany({
       where: {
@@ -74,18 +80,15 @@ exports.search = async (req, res) => {
       },
     });
 
-    if(client.length == 0){
-      res.json({status: 200, message: 'Cliente não encontrado'})
+    if(client.length < 1){
+      res.status(404).json({ message: 'Cliente não encontrado'})
     } else {
-      res.json({status: 200, client})
+      res.status(200).json({client})
     }
     
   } catch (err) {
     console.error(err);
-    res.json({
-      status: 400,
-      error: "Dados não encontrados",
-    });
+    res.status(500).end();
   }
 };
 
@@ -94,7 +97,6 @@ exports.updateOne = async (req, res) => {
     const {
       nome,
       cnpj,
-      cnpjApto,
       razaoSocial,
       endereco,
       numero,
@@ -104,8 +106,20 @@ exports.updateOne = async (req, res) => {
       emailResponsavel,
       nomeResponsavel,
       telefoneResponsavel,
-      Situacao
+      situacao
     } = req.body;
+
+    if( !nome || 
+        !cnpj ||
+        !razaoSocial || 
+        !emailResponsavel ||
+        !nomeResponsavel || 
+        !situacao){
+          return res.status(400).json({message: "Dados Inválidos"})
+    }
+    if(!id){
+      return res.status(400).json({message: "Dados Inválidos"})
+    }
 
     const id_ = parseInt(id);
 
@@ -117,7 +131,6 @@ exports.updateOne = async (req, res) => {
       data:{
         nome,
         cnpj,
-        cnpjApto,
         razaoSocial,
         endereco,
         numero,
@@ -127,37 +140,35 @@ exports.updateOne = async (req, res) => {
         emailResponsavel,
         nomeResponsavel,
         telefoneResponsavel,
-        Situacao
+        situacao
       }
     });
-    res.json({ status: 200, message: 'Cliente editado com sucesso!', clientUpdate });
+    res.status(200).json({message: 'Cliente editado com sucesso!', clientUpdate });
   }catch (err) {
     console.error(err);
-    res.json({
-      status: 400,
-      error: "Dados não editados",
-    });
-}
+    res.status(500).end();
+  }
 };
 
 
 exports.deleteOne = async (req, res) => {
   const { id } = req.params;
+  if(!id){
+    return res.status(400).json({message: "Dados Inválidos"})
+  }
   const id_ = parseInt(id);
 
   try {
     const client_delete = await prisma.client.delete({
       where: {
-        id: id_,
+        id: id_
       },
     });
 
-    res.json({ status: 204, message: 'Cliente removido com sucesso!' });
+    res.status(200).json({message: 'Cliente removido com sucesso!', client_delete});
 
   } catch (err) {
     console.error(err);
-    res.json({
-      status: 400,
-      error: "Dados não deletados",
-    })}
+    res.status(500).end();
+  }
 };
